+++ NplotCompact/PlotSurface2D.cs
@@ -530,10 +530,10 @@ namespace NPlot
 			drawables_.Add( p );
 			xAxisPositions_.Add( xp );
 			yAxisPositions_.Add( yp );
-            zPositions_.Add((float)zOrder);
+            zPositions_.Add((double)zOrder);
             // fraction is to make key unique. With 10 million plots at same z, this buggers up.. 
-            float fraction = (float)(++uniqueCounter_)/10000000.0f; 
-            ordering_.Add( (float)zOrder + fraction, drawables_.Count - 1 );
+            double fraction = (double)(++uniqueCounter_)/10000000.0f; 
+            ordering_.Add( (double)zOrder + fraction, drawables_.Count - 1 );
             
             // if p is just an IDrawable, then it can't affect the axes.
 			if ( p is IPlot )
@@ -1029,7 +1029,7 @@ namespace NPlot
 			{
 	
                 int i = (int)ordering_.GetByIndex(i_o);
-				float zOrder = (float)ordering_.GetKey( i_o );
+				double zOrder = (double)ordering_.GetKey( i_o );
 				if (zOrder > this.legendZOrder_)
 				{
 					// draw legend.
@@ -1215,9 +1215,9 @@ namespace NPlot
 			ordering_ = new SortedList();
 			for (int i = 0; i < zPositions_.Count; ++i) 
 			{
-                float zpos = Convert.ToSingle(zPositions_[i]);
-				float fraction = (float)(++uniqueCounter_) / 10000000.0f;
-				float d = zpos + fraction;
+				double zpos = Convert.ToDouble(zPositions_[i]);
+				double fraction = (double)(++uniqueCounter_) / 10000000.0f;
+				double d = zpos + fraction;
 				ordering_.Add(d, i);
 			}
 		}
